- updated touched files by last moves
