+++ src/CLI/Model/Context.h
@@ -7,7 +7,7 @@
 
 #include <string>
 #include <vector>
-#include "Core/api/TaskServiceCreator.h"
+#include "Core/utils/TaskServiceCreator.h"
 
 /*
  *  Storage of current tasks and taskService.
